Push all tags if empty string supplied as tag
It's desirable to be able to push all tags on an image, rather than
having to specify them all individually (especially if there are several
tags, which there may be).

This also mirrors the functionality of the `docker push` cli command,
which pushes all tags if no tags are specified.

This change preserves the existing default of tagging with the
`$CIRCLE_SHA1` when no tag parameter is specified.  Arguably pushing all
the tags would be a better default, but keeping as is to retain
backwards compatibility.