Add tag-after-deploy hook for better tagging
It is important to be able to add tags easily after
the docker image has been pushed (e.g. version numbers
of software installed on the image, which can only be
known after building the image, not before).

By adding a `tag-after-deploy` hook into the `publish`
job, it makes the tagging very clean in terms of config, e.g.

```
jobs:
    - docker/publish:
        tag_after_deploy: tag1,tag2
```